+++ bin/core/frameworks/index.mjs
@@ -1,13 +1,17 @@
 import { nextjsAdapter } from "./nextjs.mjs";
 import { nuxtAdapter } from "./nuxt.mjs";
 import { reactViteAdapter } from "./react-vite.mjs";
+import { sveltekitAdapter } from "./sveltekit.mjs";
 import { vueViteAdapter } from "./vue-vite.mjs";
 
 const adapters = new Map(
-  [reactViteAdapter, nextjsAdapter, vueViteAdapter, nuxtAdapter].map((adapter) => [
-    adapter.id,
-    adapter,
-  ])
+  [
+    reactViteAdapter,
+    nextjsAdapter,
+    vueViteAdapter,
+    nuxtAdapter,
+    sveltekitAdapter,
+  ].map((adapter) => [adapter.id, adapter])
 );
 
 export function getFrameworkAdapter(framework) {